+++ src/test/java/com/openggf/tests/HeadlessTestRunner.java
@@ -209,6 +209,10 @@ public class HeadlessTestRunner {
         driver.advancePlayableAnimationsOnly();
     }
 
+    public void advancePlayableFixedSlotsOnly() {
+        driver.advancePlayableFixedSlotsOnly();
+    }
+
     public void suppressFirstSidekickAnimationOnce() {
         driver.suppressFirstSidekickAnimationOnce();
     }